# docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.UpdateMeRequest:
    properties:
      name:
        type: string
    type: object
...
  model.ExternalIdentity:
    properties:
      chat_id:
        description: provider-specific (Telegram DM chat)
        type: string
      created_at:
        type: string
      display_name:
        description: shown in timeline / UI
        type: string
      external_id:
        type: string
      id:
        type: string
      provider:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  model.TeamMemberRole:
    enum:
    - team_admin
    - team_member
    type: string
    x-enum-comments:
      TeamMemberRoleAdmin: Team administrator
      TeamMemberRoleMember: Regular team member
    x-enum-varnames:
    - TeamMemberRoleAdmin
    - TeamMemberRoleMember
...
  model.User:
    properties:
      auth_provider:
        description: '"password" or "oidc" - last login method'
        type: string
      created_at:
        type: string
      email:
        description: Unique email
        type: string
      id:
        description: UUID
        type: string
      identities:
        description: loaded on demand by handlers
        items:
          $ref: '#/definitions/model.ExternalIdentity'
        type: array
      name:
        description: Display name
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        description: 'Global role: admin or user'
    type: object
  model.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-comments:
      UserRoleAdmin: Full system access
      UserRoleUser: Standard authenticated user (default)
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
  model.UserWithPermissions:
    properties:
      auth_provider:
        description: '"password" or "oidc" - last login method'
        type: string
      created_at:
        type: string
      email:
        description: Unique email
        type: string
      id:
        description: UUID
        type: string
      identities:
        description: loaded on demand by handlers
        items:
          $ref: '#/definitions/model.ExternalIdentity'
        type: array
      name:
        description: Display name
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        description: 'Global role: admin or user'
      teams:
        additionalProperties:
          $ref: '#/definitions/model.TeamMemberRole'
        description: Teams is a map of TeamID to Role (e.g. "devops" -> "team_admin")
        type: object
    type: object
...
info:
  contact: {}
  description: TokayOps API
  title: TokayOps API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and set HTTPOnly cookie
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/auth/me:
    get:
      description: Get currently authenticated user details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserWithPermissions'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get current user
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Update the current user's name. SSO users cannot change their name.
      parameters:
      - description: Profile updates
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/api.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update current user profile
      tags:
      - auth
...
  /api/auth/oidc/callback:
    get:
      description: Completes OIDC login flow, creates user if needed
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Redirect to home
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Handle OIDC callback
      tags:
      - auth
...
  /api/auth/oidc/redirect:
    get:
      description: Initiates OIDC login flow
      responses:
        "302":
          description: Redirect to OIDC provider
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Redirect to OIDC provider
      tags:
      - auth
...
swagger: "2.0"